Get changes from dalerank - building preview.

Former-commit-id: bf52c669cd4e4449e837b1090732d3b85526c979

+++ caesar.cpp
@@ -202,7 +202,7 @@ void CaesarApp::setScreenMenu()
       {
         /* temporary*/
         std::vector <fs::path> filelist = scanForMaps();
-	std::srand(std::time(0));
+	std::srand( (Uint32)std::time(0));
 	std::string file = filelist.at(std::rand()%filelist.size()).string();
 	std::cout<<"Loading map:" << file << std::endl;
         loadScenario(file);
